created a MealList component to handle output for both Categories and Favorites stacks
