+++ screens/CategoryMealsScreen.js
@@ -1,27 +1,10 @@
 import React from 'react';
-import { View, Text, FlatList, StyleSheet } from 'react-native';
 
 import { CATEGORIES, MEALS } from '../data/dummy-data';
-import MealItem from '../components/MealItem';
+import MealList from '../components/MealList';
 
 const CategoryMealsScreen = props => {
-    const renderMealItem = itemData => {
-        return <MealItem 
-            title={itemData.item.title}
-            image={itemData.item.imageUrl}
-            duration={itemData.item.duration}
-            complexity={itemData.item.complexity}
-            affordability={itemData.item.affordability}
-            onSelectMeal={() => {
-                props.navigation.navigate({
-                    routeName: 'MealDetail',
-                    params: {
-                        mealId: itemData.item.id
-                    }
-                })
-            }} 
-        />;
-    };
+    
 
     const catId = props.navigation.getParam('categoryId');
 
@@ -29,16 +12,7 @@ const CategoryMealsScreen = props => {
         meal => meal.categoryIds.indexOf(catId) >= 0
     );
 
-    return (
-        <View style={styles.screen}>
-            <FlatList 
-                data={displayedMeals} 
-                keyExtractor={(item, index) => item.id} 
-                renderItem={renderMealItem}
-                style={{ width: '100%' }}
-            />
-        </View>
-    )
+    return <MealList listData={displayedMeals} navigation={props.navigation} />;
 };
 
 CategoryMealsScreen.navigationOptions = (navigationData) => {
@@ -50,14 +24,5 @@ CategoryMealsScreen.navigationOptions = (navigationData) => {
         headerTitle: selectedCategory.title
     }
 }
-
-const styles= StyleSheet.create({
-    screen: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        padding: 15
-    }
-});
 
 export default CategoryMealsScreen;
